requirements.txt
backports.tarfile==1.2.0
bcrypt==4.3.0
cryptography==45.0.5
exceptiongroup==1.3.0
fastapi==0.116.1
fastapi-mail==1.5.0
importlib-metadata==8.0.0
inflect==7.3.1
jaraco-functools==4.3.0
jaraco.collections==5.1.0
packaging==24.2
passlib==1.7.4
pip-chill==1.0.3
platformdirs==4.2.2
python-jose==3.5.0
python-multipart==0.0.20
sqlalchemy==2.0.41
stripe==13.0.1
tomli==2.0.1
tronpy==0.6.1
uvicorn==0.35.0
web3==7.13.0
